livekit-agent/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==15.0.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.2.2
flask==3.1.0
flask-cors==5.0.0
frozenlist==1.7.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
gunicorn==24.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.10.0
livekit==1.0.12
livekit-agents==1.2.2
livekit-api==1.0.5
livekit-blingfire==1.0.0
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.2.2
livekit-protocol==1.0.4
markupsafe==3.1.1
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.97.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
pillow==11.3.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
sounddevice==0.5.2
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
